+++ autoprotocol_utilities/__init__.py
@@ -9,3 +9,6 @@ from .resource_helpers import ResourceIDs, oligo_scale_default, \
     return_dispense_media, return_agar_plates, ref_kit_container, \
     oligo_dilution_table
 from .thermocycle_helpers import melt_curve, thermocycle_ramp
+from .bio_calculators import dna_mass_to_mole, dna_mole_to_mass, \
+    molar_to_mass_conc, mass_conc_to_molar, ligation_insert_ng, \
+    ligation_insert_volume, ligation_insert_amount